re-update company và admin
